+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/HistoryDelegateTest.kt
@@ -174,7 +174,7 @@ class HistoryDelegateTest {
             fail()
         }
 
-        override suspend fun runMaintenance() {
+        override suspend fun runMaintenance(dbSizeLimit: UInt) {
             fail()
         }